remove pcap rotate option. Raise condition between /opt/pcap-filtered and /opt/pcap when disk space is less than watermark

+++ files/bin/pcapture-filter.sh
@@ -60,7 +60,7 @@ fi
 
 for FILE in $(tail -n +${NUM} ${DSTDIR}/${LOG_ALLFILES}) ; do
 	if [ ! -e ${DSTDIR}/${FILE} ] ; then
-		#unxz -k -c /opt/pcap/20190401-000403_300.ignored.pcap.xz | tcpdump -n -r -
+		# unxz -k -c /opt/pcap/20190401-000403_300.ignored.pcap.xz | tcpdump -n -r -
 		# Original filename like: 20190423-201634_300.ignored.pcap.xz
 		# New file name (filtered): 20190423-201634_300.filtered.pcap.xz
 		newFILE="$(echo ${FILE} | cut -d '.' -f1)"
@@ -73,7 +73,7 @@ for FILE in $(tail -n +${NUM} ${DSTDIR}/${LOG_ALLFILES}) ; do
 				${XZ} ${DSTDIR}/${newFILE}.temp
 				# After it's compressed we renamed it (to avoid to transfer a file in the middle of compression process)
 				mv ${DSTDIR}/${newFILE}.temp.xz ${DSTDIR}/${newFILE}.filtered.pcap.xz
-			else
+			elif [ -f ${DSTDIR}/${newFILE}.temp ] ; then
 				rm ${DSTDIR}/${newFILE}.temp
 			fi
 			echo ${FILE} > ${DSTDIR}/${LOG_LASTFILE}
